individually allowing users to assign chores through chore cards
